+++ Demo.AspNetCore.ServerTiming/Program.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.AspNetCore.Builder;
+using Demo.AspNetCore.ServerTiming;
+
+var builder = WebApplication.CreateBuilder(args);
+
+var startup = new Startup(builder.Configuration);
+
+startup.ConfigureServices(builder.Services);
+
+var app = builder.Build();
+
+startup.Configure(app, app.Environment);
+
+app.Run();