+++ DocFx.AspNetCore.ServerTiming/Program.cs
@@ -1,0 +1,9 @@
+using Microsoft.AspNetCore.Builder;
+
+var builder = WebApplication.CreateBuilder(args);
+var app = builder.Build();
+
+app.UseDefaultFiles()
+    .UseStaticFiles();
+
+app.Run();