+++ Lib.AspNetCore.ServerTiming/ServerTimingUtility.cs
@@ -39,6 +39,29 @@ namespace Lib.AspNetCore.ServerTiming
         /// <summary>
         /// Time an async task.
         /// </summary>
+        /// <param name="task">The <see cref="Task"/> to time.</param>
+        /// <param name="serverTiming">The <see cref="IServerTiming"/> to add metric to.</param>
+        /// <param name="metricName">Optional, metric name, if passed will override any caller name.</param>
+        /// <param name="functionName">Optional, populated compile-time with the name of the calling function.</param>
+        /// <param name="filePath">Optional, populated compile-time with the path to the calling file.</param>
+        /// <param name="lineNumber">Optional, populated compile-time with line number in the calling file.</param>
+        /// <returns></returns>
+        public static async Task TimeTask(this IServerTiming serverTiming,
+            Task task,
+            string metricName = null,
+            [CallerMemberName] string functionName = null,
+            [CallerFilePath] string filePath = null,
+            [CallerLineNumber] int lineNumber = 0)
+        {
+            using (serverTiming.TimeAction(metricName, functionName, filePath, lineNumber))
+            {
+                await task;
+            }
+        }
+
+        /// <summary>
+        /// Time an async task.
+        /// </summary>
         /// <typeparam name="T">Type of the task.</typeparam>
         /// <param name="task">The <see cref="Task"/> to time.</param>
         /// <param name="serverTiming">The <see cref="IServerTiming"/> to add metric to.</param>
@@ -61,7 +84,7 @@ namespace Lib.AspNetCore.ServerTiming
         }
 
         /// <summary>
-        /// Add a metric to the timing, if present.
+        /// Add a timed metric to the timing, if present.
         /// </summary>
         /// <param name="serverTiming">The <see cref="IServerTiming"/> to add metric to.</param>
         /// <param name="duration">The duration to log in ms.</param>
@@ -81,6 +104,32 @@ namespace Lib.AspNetCore.ServerTiming
             }
 
             ServerTimingMetric metric = new ServerTimingMetric(metricName ?? FormatCallerName(functionName, filePath, lineNumber), duration);
+
+            serverTiming.Metrics.Add(metric);
+        }
+
+        /// <summary>
+        /// Add an untimed metric to the timing, if present.
+        /// </summary>
+        /// <param name="serverTiming">The <see cref="IServerTiming"/> to add metric to.</param>
+        /// <param name="metricName">The name of the metric to log.</param>
+        /// <param name="description">The description of the metric</param>
+        /// <param name="functionName">Optional, populated compile-time with the name of the calling function</param>
+        /// <param name="filePath">Optional, populated compile-time with the path to the calling file</param>
+        /// <param name="lineNumber">Optional, populated compile-time with line number in the calling file</param>
+        public static void AddMetric(this IServerTiming serverTiming,
+            string metricName = null,
+            string description = null,
+            [CallerMemberName] string functionName = null,
+            [CallerFilePath] string filePath = null,
+            [CallerLineNumber] int lineNumber = 0)
+        {
+            if (serverTiming is null)
+            {
+                return;
+            }
+
+            ServerTimingMetric metric = new ServerTimingMetric(metricName ?? FormatCallerName(functionName, filePath, lineNumber), description);
 
             serverTiming.Metrics.Add(metric);
         }