+++ Test.AspNetCore.ServerTiming/ServerTimingServerTests.cs
@@ -1,8 +1,13 @@
 ﻿using System.Net.Http;
 using System.Threading.Tasks;
 using System.Collections.Generic;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.TestHost;
+using Microsoft.Extensions.DependencyInjection;
+using Lib.AspNetCore.ServerTiming.Filters;
+using Lib.AspNetCore.ServerTiming.Http.Headers;
+using Moq;
 using Xunit;
 using Test.AspNetCore.ServerTiming.Infrastructure;
 
@@ -16,12 +21,23 @@ namespace Test.AspNetCore.ServerTiming
         #endregion
 
         #region Prepare SUT
-        private TestServer PrepareTestServer()
+        private TestServer PrepareTestServer(List<IServerTimingMetricFilter> filters = null)
         {
-            IWebHostBuilder webHostBuilder = new WebHostBuilder()
-                .UseStartup<ServerTimingServerStartup>();
+            IWebHostBuilder webHostBuilder = new WebHostBuilder().ConfigureServices(services =>
+            {
+                services.AddSingleton<IStartup>(new ServerTimingServerStartup(filters ?? new List<IServerTimingMetricFilter>()));
+            });
 
             return new TestServer(webHostBuilder);
+        }
+
+        private Mock<IServerTimingMetricFilter> PrepareServerTimingMetricFilterMock(bool onServerTimingHeaderPreparationResult = true)
+        {
+            Mock<IServerTimingMetricFilter> serverTimingMetricFilterMock = new Mock<IServerTimingMetricFilter>();
+
+            serverTimingMetricFilterMock.Setup(m => m.OnServerTimingHeaderPreparation(It.IsAny<HttpContext>(), It.IsAny<ICollection<ServerTimingMetric>>())).Returns(onServerTimingHeaderPreparationResult);
+
+            return serverTimingMetricFilterMock;
         }
         #endregion
 
@@ -56,7 +72,7 @@ namespace Test.AspNetCore.ServerTiming
             }
         }
 
-#if NETCOREAPP3_0
+#if !NETCOREAPP2_1 && !NET461
         [Fact]
         public async Task Request_AllowsTrailers_ReturnsResponseWithServerTimingTrailer()
         {
@@ -74,6 +90,50 @@ namespace Test.AspNetCore.ServerTiming
             }
         }
 #endif
+
+        [Fact]
+        public async Task Request_TwoFiltersRegistered_FiltersNotShortCircuiting_AllFiltersRun()
+        {
+            Mock<IServerTimingMetricFilter> firstServerTimingMetricFilter = PrepareServerTimingMetricFilterMock();
+            Mock<IServerTimingMetricFilter> secondServerTimingMetricFilter = PrepareServerTimingMetricFilterMock();
+
+            using (TestServer server = PrepareTestServer(new List<IServerTimingMetricFilter>
+            {
+                firstServerTimingMetricFilter.Object,
+                secondServerTimingMetricFilter.Object
+            }))
+            {
+                using (HttpClient client = server.CreateClient())
+                {
+                    HttpResponseMessage response = await client.GetAsync("/");
+
+                    firstServerTimingMetricFilter.Verify(m => m.OnServerTimingHeaderPreparation(It.IsAny<HttpContext>(), It.IsAny<ICollection<ServerTimingMetric>>()), Times.Once);
+                    secondServerTimingMetricFilter.Verify(m => m.OnServerTimingHeaderPreparation(It.IsAny<HttpContext>(), It.IsAny<ICollection<ServerTimingMetric>>()), Times.Once);
+                }
+            }
+        }
+
+        [Fact]
+        public async Task Request_TwoFiltersRegistered_FirstFilterShortCircuiting_OnlyFirstFiltersRun()
+        {
+            Mock<IServerTimingMetricFilter> firstServerTimingMetricFilter = PrepareServerTimingMetricFilterMock(false);
+            Mock<IServerTimingMetricFilter> secondServerTimingMetricFilter = PrepareServerTimingMetricFilterMock();
+
+            using (TestServer server = PrepareTestServer(new List<IServerTimingMetricFilter>
+            {
+                firstServerTimingMetricFilter.Object,
+                secondServerTimingMetricFilter.Object
+            }))
+            {
+                using (HttpClient client = server.CreateClient())
+                {
+                    HttpResponseMessage response = await client.GetAsync("/");
+
+                    firstServerTimingMetricFilter.Verify(m => m.OnServerTimingHeaderPreparation(It.IsAny<HttpContext>(), It.IsAny<ICollection<ServerTimingMetric>>()), Times.Once);
+                    secondServerTimingMetricFilter.Verify(m => m.OnServerTimingHeaderPreparation(It.IsAny<HttpContext>(), It.IsAny<ICollection<ServerTimingMetric>>()), Times.Never);
+                }
+            }
+        }
         #endregion
     }
 }